+++ CHANGES.md
@@ -2,9 +2,11 @@
 
 # Changes for Fractal core
 
-# Next v2.0.0
+# Next
 
 ## Done
+
+# v2.0.0
 
 - Add stateOf, toIt, toChild, nest, unnest, nestAll and unnestAll to input helpers and curry them
 - Group all input helpers to inputs.ts